Revert "add Live styles and components"
This reverts commit 0378c7e5a2c48aa4dedd748935cb9c170f409b7a.
